cookie_store: fix missing raw cookie elements
Some raw cookie elements (domain & path) are lost
when adding a raw cookie to a cookie store, so
retrieving back the raw cookie from the store
results in information loss.

Fix this by ensuring the fields are added to the
raw cookie builder in try_from_raw_cookie().

We also fix two tests with the missing values.